move splitOverlappingTermsIntoSegments and splitTermIntoSegments into their own file
