+++ packages/react-tei/src/unitex/splitOverlappingTermsIntoSegments.ts
@@ -1,0 +1,195 @@
+import { propagateGroupsToSubTerm } from "./propagateGroupsToSubTerm";
+import type { NestedTerm, TermWithPosition } from "./types";
+
+type SubTermAtPosition = {
+	subTerm: NestedTerm;
+	start: number;
+	end: number;
+	fromTermGroups: string[];
+};
+
+// Collect all boundary points from overlapping terms and their subTerms
+const collectBoundaries = (overlappingGroup: TermWithPosition[]): number[] => {
+	const boundaries = new Set<number>();
+
+	for (const t of overlappingGroup) {
+		boundaries.add(t.start);
+		boundaries.add(t.end);
+
+		// Add boundaries from subTerms
+		if (t.term.subTerms) {
+			let subPos = t.start;
+			for (const st of t.term.subTerms) {
+				boundaries.add(subPos);
+				boundaries.add(subPos + st.term.length);
+				subPos += st.term.length;
+			}
+		}
+	}
+
+	return Array.from(boundaries).sort((a, b) => a - b);
+};
+
+// Build subTerms position map from overlapping terms
+const buildSubTermsPositionMap = (
+	overlappingGroup: TermWithPosition[],
+): SubTermAtPosition[] => {
+	return overlappingGroup.flatMap((t) => {
+		if (!t.term.subTerms) return [];
+
+		let subPos = t.start;
+		return t.term.subTerms.map((st) => {
+			const result = {
+				subTerm: st,
+				start: subPos,
+				end: subPos + st.term.length,
+				fromTermGroups: t.term.groups,
+			};
+			subPos += st.term.length;
+			return result;
+		});
+	});
+};
+
+// Compute source term from covering terms
+// Returns slugified term if all non-artificial terms have the same term, null otherwise
+const computeSourceTerm = (
+	coveringTerms: TermWithPosition[],
+): string | null => {
+	const nonArtificialTerms = coveringTerms.filter((t) => !t.term.artificial);
+
+	if (nonArtificialTerms.length === 0) return null;
+
+	const uniqueTermValues = [
+		...new Set(nonArtificialTerms.map((t) => t.term.term)),
+	];
+
+	if (uniqueTermValues.length === 1) {
+		return uniqueTermValues[0]!;
+	}
+
+	return null;
+};
+
+// Compute groups filtering out artificial terms
+const computeNonArtificialGroups = (
+	coveringTerms: TermWithPosition[],
+	parentGroups: string[],
+): string[] => {
+	// Filter out groups from artificial terms
+	const nonArtificialGroups = coveringTerms
+		.filter((t) => !t.term.artificial)
+		.flatMap((t) => t.term.groups);
+
+	return [...new Set([...parentGroups, ...nonArtificialGroups])].sort();
+};
+
+// Create a segment from boundary positions
+const createSegmentFromBoundary = (
+	segStart: number,
+	segEnd: number,
+	containerTerm: string,
+	overlappingGroup: TermWithPosition[],
+	allTerms: TermWithPosition[],
+	subTermsByPosition: SubTermAtPosition[],
+	parentGroups: string[],
+): NestedTerm | null => {
+	const segmentText = containerTerm.slice(segStart, segEnd);
+	if (segmentText === "") return null;
+
+	// Find which terms from the overlapping group cover this segment
+	const coveringTermsFromGroup = overlappingGroup.filter(
+		(t) => t.start <= segStart && t.end >= segEnd,
+	);
+
+	// Find ALL terms that cover this segment (for group computation)
+	const allCoveringTerms = allTerms.filter(
+		(t) => t.start <= segStart && t.end >= segEnd,
+	);
+
+	// Find subTerms that exactly match this segment
+	const matchingSubTerms = subTermsByPosition.filter(
+		(st) => st.start === segStart && st.end === segEnd,
+	);
+
+	if (matchingSubTerms.length > 0) {
+		// This segment exactly matches one or more subTerms - merge their properties
+		const baseSubTerm = matchingSubTerms[0]!.subTerm;
+
+		// For artificial subTerms (filler text), only use parentGroups
+		// For non-artificial subTerms, merge groups from non-artificial covering terms
+		const isArtificialSegment = baseSubTerm.artificial;
+		const allGroups = isArtificialSegment
+			? computeNonArtificialGroups(allCoveringTerms, [])
+			: computeNonArtificialGroups(allCoveringTerms, parentGroups);
+
+		// Compute source term (using only the overlapping group for this)
+		const sourceTerm = computeSourceTerm(coveringTermsFromGroup);
+
+		// Recursively propagate groups to nested subTerms if any
+		let nestedSubTerms = baseSubTerm.subTerms;
+		if (nestedSubTerms) {
+			nestedSubTerms = nestedSubTerms.map((st) =>
+				propagateGroupsToSubTerm(st, allGroups),
+			);
+		}
+
+		return {
+			term: segmentText,
+			groups: allGroups,
+			sourceTerm,
+			...(baseSubTerm.artificial && { artificial: true }),
+			...(nestedSubTerms &&
+				nestedSubTerms.length > 0 && { subTerms: nestedSubTerms }),
+		};
+	}
+
+	if (coveringTermsFromGroup.length === 0) {
+		// No covering term from the overlapping group - this is filler text
+		return {
+			term: segmentText,
+			groups: parentGroups,
+			artificial: true,
+		};
+	}
+
+	// Covered by terms but no exact subTerm match - create segment from covering terms
+	const groups = computeNonArtificialGroups(allCoveringTerms, parentGroups);
+	const sourceTerm = computeSourceTerm(coveringTermsFromGroup);
+
+	return {
+		term: segmentText,
+		groups,
+		sourceTerm,
+	};
+};
+
+// Process multiple overlapping terms into segments
+export const splitOverlappingTermsIntoSegments = (
+	overlappingGroup: TermWithPosition[],
+	allTerms: TermWithPosition[],
+	containerTerm: string,
+	parentGroups: string[],
+): NestedTerm[] => {
+	const sortedBoundaries = collectBoundaries(overlappingGroup);
+	const subTermsByPosition = buildSubTermsPositionMap(overlappingGroup);
+
+	// Build segments from consecutive boundary pairs
+	const segments: NestedTerm[] = [];
+	for (let k = 0; k < sortedBoundaries.length - 1; k++) {
+		const segment = createSegmentFromBoundary(
+			sortedBoundaries[k]!,
+			sortedBoundaries[k + 1]!,
+			containerTerm,
+			overlappingGroup,
+			allTerms,
+			subTermsByPosition,
+			parentGroups,
+		);
+		if (segment) {
+			segments.push(segment);
+		}
+	}
+
+	return segments;
+};
